moved cli to seperate folder
