+++ cli/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "uptest"
+name = "uptest-cli"
 version = "0.1.2"
 edition = "2021"
 license = "MIT"
@@ -7,18 +7,10 @@ homepage = "https://github.com/uptest-sc/uptest/"
 description = "substrate rUntime uPgrade TESTing suit"
 readme = "README.md"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[workspace]
-members = [
-	"libuptest",
-	"cli",
-	"examples"
-]
-
 
 [dependencies]
 clap = { version = "4.1.4", features=["derive"]}
-libuptest = { path = "libuptest/", version = "0.1.2", features = ["metadatadecode"]}
+libuptest = { path = "../libuptest/", version = "0.1.2", features = ["metadatadecode"]}
 tokio = { version = "1.27", features = ["full"] }
 hex = { version = "0.4.3", features = ["alloc"]}
 anyhow = "1.0.69"
