+++ cli/src/cli.rs
